Rename classes for more consistency
