+++ bukkit/src/main/java/me/neznamy/tab/platforms/bukkit/scoreboard/viaversion/ViaScoreboard116.java
@@ -1,4 +1,4 @@
-package me.neznamy.tab.platforms.bukkit.scoreboard;
+package me.neznamy.tab.platforms.bukkit.scoreboard.viaversion;
 
 import com.viaversion.viaversion.api.protocol.Protocol;
 import com.viaversion.viaversion.api.protocol.packet.PacketType;
@@ -14,18 +14,18 @@ import me.neznamy.tab.platforms.bukkit.BukkitTabPlayer;
  * Scoreboard implementation using ViaVersion packets
  * to unlock +1.16 player features on pre 1.16 servers.
  */
-public class ViaScoreboard16 extends ViaScoreboard13 {
+public class ViaScoreboard116 extends ViaScoreboard113 {
 
     /**
      * Constructs new instance with given player.
      *
      * @param player          Player this scoreboard will belong to
      */
-    public ViaScoreboard16(@NonNull BukkitTabPlayer player) {
+    public ViaScoreboard116(@NonNull BukkitTabPlayer player) {
         this(player, Protocol1_15_2To1_16.class, ClientboundPackets1_16.SET_DISPLAY_OBJECTIVE, ClientboundPackets1_16.SET_OBJECTIVE, ClientboundPackets1_16.SET_SCORE, ClientboundPackets1_16.SET_PLAYER_TEAM);
     }
 
-    protected ViaScoreboard16(@NonNull BukkitTabPlayer player, @NonNull Class<? extends Protocol> protocol, @NonNull PacketType setDisplayObjective, @NonNull PacketType setObjective, @NonNull PacketType setScore, @NonNull PacketType setPlayerTeam) {
+    protected ViaScoreboard116(@NonNull BukkitTabPlayer player, @NonNull Class<? extends Protocol> protocol, @NonNull PacketType setDisplayObjective, @NonNull PacketType setObjective, @NonNull PacketType setScore, @NonNull PacketType setPlayerTeam) {
         super(player, protocol, setDisplayObjective, setObjective, setScore, setPlayerTeam);
     }
