+++ bukkit/src/main/java/me/neznamy/tab/platforms/bukkit/tablist/viaversion/ViaTabList1214.java
@@ -1,4 +1,4 @@
-package me.neznamy.tab.platforms.bukkit.tablist;
+package me.neznamy.tab.platforms.bukkit.tablist.viaversion;
 
 import com.viaversion.viaversion.api.protocol.packet.PacketWrapper;
 import com.viaversion.viaversion.api.type.Types;
